requirements.txt
# Core app
streamlit>=1.33,<2

# Data + parsing
pandas>=2.1,<3
numpy>=1.26,<3
lxml>=4.9,<6    # for XML parsing in your hdb* modules

# DOCX generation
python-docx>=1.0.0

# Optional: if you read/save parquet/arrow anywhere; remove if unused
pyarrow>=10,<18

# Optional: if any of your parsers fetch remote files (not visible in streamlit_app.py)
requests>=2.31,<3